+++ crates/ntx-builder/src/builder/mod.rs
@@ -7,7 +7,7 @@ use anyhow::Context;
 use futures::TryStreamExt;
 use miden_node_proto::domain::account::NetworkAccountPrefix;
 use miden_node_utils::ErrorReport;
-use miden_remote_prover_client::remote_prover::tx_prover::RemoteTransactionProver;
+use miden_remote_prover::remote_prover::tx_prover::RemoteTransactionProver;
 use tokio::sync::Barrier;
 use tokio::time;
 use url::Url;