+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - [BREAKING] De-duplicate storage of code in DB (no-migration) ([#1083](https://github.com/0xMiden/miden-node/issue/#1083)).
 - [BREAKING] RPC accept header format changed from `application/miden.vnd+grpc.<version>` to `application/vnd.miden; version=<version>` ([#1084](https://github.com/0xMiden/miden-node/pull/1084)).
 - [BREAKING] Integrated `FeeParameters` into block headers. ([#1122](https://github.com/0xMiden/miden-node/pull/1122)).
+- [BREAKING] Merged `miden-remote-prover` and `miden-remote-prover-client` into a single `miden-remote-prover` crate. ([#1163](https://github.com/0xMiden/miden-node/pull/1163)).
 
 ### Fixes